+++ NEWS.md
@@ -17,7 +17,8 @@ argument, which, if `TRUE`, will label any Pareto `k` points greater than
 1/2 by the index number of the corresponding observation. The plot method 
 also now warns about `Inf`/`NA`/`NaN` values of `k` that are not shown in 
 the plot. 
-* Allow setting number of cores using `options(loo.cores = NUMBER)`.
+* `compare` now returns model weights.
+* Allow setting number of cores using `options(loo.cores = NUMBER)`. 
 
 # loo 0.1.2 
 * Updates names in package to reflect name changes in the accompanying